backend/requirements.txt
# Core data
pandas==2.2.3
numpy==1.26.4

# APIs financieras
yfinance==0.2.50

# Supabase
supabase==2.10.0
python-dotenv==1.0.1

# Indicadores técnicos
ta==0.11.0

# Utilidades
requests==2.32.3
python-dateutil==2.9.0
